Add files via upload
Code for full swerve, our one-module test bot, and to configure the PID constants for the turn motor of a module. This is just after one write through and there are some inefficiencies/eccentricities, so they will need to be further edited for consistency and to get more optimized results. For the PID config, it would be nice to use network tables to create graphs of the motor position in Smart Dashboard, but Ethan will need to do that.