Add files via upload
With some minor edits to the standard full swerve code I committed a few weeks ago, the joystick inputs can be translated so that forward, right, etc. makes the robot move forward, right, etc. with respect to the drivers rather than with respect to itself, regardless of the robot's actual heading. What we end up using will be up to the drivers, but both had potential pros and cons at a glance.
Standard seems like it would be better for precise aiming while more restrictive when moving around the field.
Field-oriented seems like it would be much smoother for general field movements, but require more practice to get good at fine alignment. Because the alignment issue could be mitigated by vision alignment and extra practice, this is the mode I'm leaning towards. But I'm not the one who has to drive it :P