Finalized Initial Rewritten Swerve
Well that was a fun commit name. I added the actual motor control and swerve math. The swerve math was just copied from the swerve derivation in github, so please check it against some other sources.
I decided to instantiate the DriveTrain class as drive and change the drive method to be called move. Hopefully all the names make at least a little sense. There aren't that many comments but Ethan should understand all of it - just make sure to add explanation comments whenever someone questions something.

It all works in sim but that doesn't mean much so lots of debugging will still need to happen. I have pretty spotty service in Ridgeview but should be able to get texts every hour or so. Good luck tomorrow!

p.s. I know that's a lot of information but keep in mind that this is the first meeting I've missed all year